Update supported board marcos part3
- Update for compatibility of "amb_ard_pin_check_type" and "amb_ard_pin_check_fun"
- Update coding style

+++ Arduino_package/hardware/libraries/WiFi/src/WiFiClient.h
@@ -39,7 +39,7 @@ class WiFiClient : public Client {
         using Print::write;
 
     private:
-        int _sock;
+        uint8_t _sock;
         ServerDrv clientdrv;
         bool _is_connected;
         uint8_t data[DATA_LENTH];
